Initial Routing added, file structure updated
