+++ package.json
@@ -15,11 +15,13 @@
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
     "react-parallax-tilt": "^1.7.266",
+    "react-router": "^7.0.2",
     "react-scroll": "^1.9.0",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
+    "@types/node": "^22.10.2",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/styled-components": "^5.1.34",
