+++ src/app/App.tsx
@@ -1,0 +1,9 @@
+import { Routing } from '../common/router/Routing.tsx'
+
+export function App() {
+  return (
+    <>
+      <Routing/>
+    </>
+  )
+}
