+++ src/components/CardPage/CardPage.tsx
@@ -1,0 +1,8 @@
+
+export const CardPage = () => {
+  return (
+    <div>
+
+    </div>
+  )
+}
