--- src/components/FlexWrapper.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-
-export const FlexWrapper = styled.div`
-  display: flex;
-  flex-direction:row;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: nowrap;
-  height: 100%;
-`
