+++ src/components/Header/HeaderStyled.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+export const HeaderStyled = styled.header`
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    background: rgba(31, 31, 32, 0.9);
+    padding: 10px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    z-index: 9;
+`
