+++ src/components/Icons/BaskedButton.tsx
@@ -1,4 +1,4 @@
-import { IconWrapper } from '../IconWrapper.tsx'
+import { IconWrapper } from './IconWrapperStyled.ts'
 import { FaCartShopping } from 'react-icons/fa6'
 import { ICON_CONFIG } from './IconConfig.ts'
