+++ src/components/Icons/HomeButton.tsx
@@ -1,5 +1,5 @@
 import { FaShop } from 'react-icons/fa6'
-import { IconWrapper } from '../IconWrapper.tsx'
+import { IconWrapper } from './IconWrapperStyled.ts'
 import { animateScroll as scroll } from 'react-scroll'
 import { ICON_CONFIG } from './IconConfig.ts'
