+++ src/components/Icons/LikeButton.tsx
@@ -1,4 +1,4 @@
-import { IconWrapper } from '../IconWrapper.tsx'
+import { IconWrapper } from './IconWrapperStyled.ts'
 import { MdFavorite } from 'react-icons/md'
 import { ICON_CONFIG } from './IconConfig.ts'
