+++ src/components/SearchBar/SearchBarStyled.ts
