--- src/layout/footer/FooterStyled.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-import { Theme } from '../../styles/Theme.ts'
-
-export const FooterStyled = styled.footer`
-  //background-color: ${Theme.colors.primaryBg};
-  padding: 20px 0;
-  //position: relative;
-`
