--- src/layout/header/HeaderStyled.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const HeaderStyled = styled.header`
-  background: rgba(31, 31, 32, 0.9);
-  padding: 10px;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 9;
-`
