+++ src/main.tsx
@@ -1,11 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './App.tsx'
-import { GlobalStyled } from './styles/GlobalStyled.tsx'
+import { App } from './app/App.tsx'
+import { GlobalStyled } from '@styles/GlobalStyled.ts'
+import { BrowserRouter } from 'react-router'
 
 createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <GlobalStyled />
-    <App />
-  </StrictMode>
+  <BrowserRouter>
+    <StrictMode>
+      <GlobalStyled />
+      <App />
+    </StrictMode>
+  </BrowserRouter>,
 )
