+++ vite.config.ts
@@ -1,7 +1,18 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react-swc'
+import path from 'path';
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+      '@assets': path.resolve(__dirname, './src/assets'),
+      '@components': path.resolve(__dirname, './src/components'),
+      '@layout': path.resolve(__dirname, './src/layout'),
+      '@styles': path.resolve(__dirname, './src/styles'),
+      '@common': path.resolve(__dirname, './src/common'),
+    },
+  },
 })
